Enable app-wide error handling through error pages
App errors are now handled using an error middleware so that the web app
can fail gracefully.

Errors are now displayed in dedicated error pages with a redirect button
to the homepage for better user experience.

A catch-all route was added to the web app, to handle navigation to
routes that don't exist.